Fix hasHash() call when moving a tree between repos
Just currently playing around with tedit (great work btw!) and working on the ability to move things in the tree via drag-and-drop .  When attempting to move a tree between repos, I ran into an error.  It appears `hasHash()` is being called with "tree" as the first arg in this instance, but it looks like it's expecting only 2 args with the hash as the first argument.  Removing the "tree" appears to fix it.